# app/Http/Controllers/AdminProposalController.php
<?php

declare(strict_types=1);

namespace App\Http\Controllers;

use App\Constants\PaginationConstants;
use App\Enums\ProposalStatus;
use App\Events\ProposalStatusChanged;
use App\Exceptions\UnauthorizedException;
use App\Helpers\ApiResponse;
use App\Http\Requests\UpdateProposalStatusRequest;
use App\Http\Resources\ProposalResource;
use App\Models\Proposal;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;
use Illuminate\Support\Facades\Log;
use OpenApi\Attributes as OA;

/**
 * Controller for admin proposal management.
 */
#[OA\Tag(name: "Admin")]
class AdminProposalController extends Controller
{
    /**
     * Display a listing of all proposals for admin.
     */
    #[OA\Get(
        path: "/admin/proposals",
        summary: "List all proposals (Admin only)",
        description: "Retrieves all proposals with filtering options. Only accessible by admin users. Includes additional filters like user_id.",
        tags: ["Admin"],
        security: [["sanctum" => []]],
        parameters: [
            new OA\Parameter(
                name: "search",
                in: "query",
                description: "Search proposals by title",
                required: false,
                schema: new OA\Schema(type: "string", example: "Laravel")
            ),
            new OA\Parameter(
                name: "tags",
                in: "query",
                description: "Filter by tag IDs (comma-separated)",
                required: false,
                schema: new OA\Schema(type: "string", example: "1,2,3")
            ),
            new OA\Parameter(
                name: "status",
                in: "query",
                description: "Filter by status",
                required: false,
                schema: new OA\Schema(type: "string", enum: ["pending", "approved", "rejected"], example: "pending")
            ),
            new OA\Parameter(
                name: "user_id",
                in: "query",
                description: "Filter by user ID",
                required: false,
                schema: new OA\Schema(type: "integer", example: 1)
            ),
            new OA\Parameter(
                name: "page",
                in: "query",
                description: "Page number",
                required: false,
                schema: new OA\Schema(type: "integer", example: 1)
            ),
            new OA\Parameter(
                name: "per_page",
                in: "query",
                description: "Items per page",
                required: false,
                schema: new OA\Schema(type: "integer", example: 15)
            ),
        ],
        responses: [
            new OA\Response(
                response: 200,
                description: "Proposals retrieved successfully",
                content: new OA\JsonContent(
                    properties: [
                        new OA\Property(property: "status", type: "string", example: "success"),
                        new OA\Property(property: "message", type: "string", example: "Proposals retrieved successfully"),
                        new OA\Property(
                            property: "data",
                            type: "object",
                            properties: [
                                new OA\Property(
                                    property: "proposals",
                                    type: "array",
                                    items: new OA\Items(ref: "#/components/schemas/Proposal")
                                ),
                                new OA\Property(
                                    property: "pagination",
                                    type: "object",
                                    properties: [
                                        new OA\Property(property: "current_page", type: "integer", example: 1),
                                        new OA\Property(property: "last_page", type: "integer", example: 5),
                                        new OA\Property(property: "per_page", type: "integer", example: 15),
                                        new OA\Property(property: "total", type: "integer", example: 75),
                                    ]
                                ),
                            ]
                        ),
                    ]
                )
            ),
            new OA\Response(response: 401, description: "Unauthenticated"),
            new OA\Response(response: 403, description: "Forbidden - Admin only"),
            new OA\Response(response: 500, description: "Server error"),
        ]
    )]
    public function index(Request $request): JsonResponse
    {
        try {
            if (! $request->user()->isAdmin()) {
                throw new UnauthorizedException;
            }

            $query = Proposal::with(['user', 'tags', 'reviews']);

            // Search by title
            if ($request->filled('search')) {
                $query->searchByTitle($request->string('search')->toString());
            }

            // Filter by tags
            if ($request->filled('tags')) {
                $tagIds = is_array($request->tags) ? $request->tags : explode(',', (string) $request->tags);
                $tagIds = array_map('intval', array_filter($tagIds));
                if (count($tagIds) > 0) {
                    $query->byTags($tagIds);
                }
            }

            // Filter by status
            if ($request->filled('status')) {
                $status = $request->string('status')->toString();
                if (in_array($status, ProposalStatus::values(), true)) {
                    $query->byStatus($status);
                }
            }

            // Filter by user
            if ($request->filled('user_id')) {
                $query->byUser((int) $request->integer('user_id'));
            }

            $perPage = min(
                max((int) $request->get('per_page', PaginationConstants::DEFAULT_PER_PAGE), PaginationConstants::MIN_PER_PAGE),
                PaginationConstants::MAX_PER_PAGE
            );

            $proposals = $query->latest()->paginate($perPage);

            return ApiResponse::success(
                'Proposals retrieved successfully',
                [
                    'proposals' => ProposalResource::collection($proposals->items()),
                    'pagination' => [
                        'current_page' => $proposals->currentPage(),
                        'last_page' => $proposals->lastPage(),
                        'per_page' => $proposals->perPage(),
                        'total' => $proposals->total(),
                    ],
                ]
            );
        } catch (UnauthorizedException $e) {
            return ApiResponse::error($e->getMessage(), $e->getCode());
        } catch (\Exception $e) {
            Log::error('Error retrieving admin proposals', [
                'error' => $e->getMessage(),
                'trace' => $e->getTraceAsString(),
            ]);

            return ApiResponse::error('Failed to retrieve proposals', 500);
        }
    }

    /**
     * Update the proposal status.
     */
    #[OA\Patch(
        path: "/admin/proposals/{id}/status",
        summary: "Update proposal status (Admin only)",
        description: "Updates the status of a proposal. Only accessible by admin users. Triggers real-time broadcast event.",
        tags: ["Admin"],
        security: [["sanctum" => []]],
        parameters: [
            new OA\Parameter(
                name: "id",
                in: "path",
                required: true,
                description: "Proposal ID",
                schema: new OA\Schema(type: "integer", example: 1)
            ),
        ],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                required: ["status"],
                properties: [
                    new OA\Property(property: "status", type: "string", enum: ["pending", "approved", "rejected"], example: "approved", description: "New proposal status"),
                ]
            )
        ),
        responses: [
            new OA\Response(
                response: 200,
                description: "Proposal status updated successfully",
                content: new OA\JsonContent(
                    properties: [
                        new OA\Property(property: "status", type: "string", example: "success"),
                        new OA\Property(property: "message", type: "string", example: "Proposal status updated successfully"),
                        new OA\Property(
                            property: "data",
                            type: "object",
                            properties: [
                                new OA\Property(property: "proposal", ref: "#/components/schemas/Proposal"),
                            ]
                        ),
                    ]
                )
            ),
            new OA\Response(response: 401, description: "Unauthenticated"),
            new OA\Response(response: 403, description: "Forbidden - Admin only"),
            new OA\Response(response: 404, description: "Proposal not found"),
            new OA\Response(response: 422, description: "Validation error"),
            new OA\Response(response: 500, description: "Server error"),
        ]
    )]
    public function updateStatus(UpdateProposalStatusRequest $request, Proposal $proposal): JsonResponse
    {
        try {
            DB::beginTransaction();

            // Get old status as string (status is cast to ProposalStatus enum)
            $oldStatus = $proposal->status instanceof ProposalStatus 
                ? $proposal->status->value 
                : (string) $proposal->status;
            $status = ProposalStatus::from($request->string('status')->toString());

            $proposal->update([
                'status' => $status->value,
            ]);

            $proposal->load(['user', 'tags']);

            DB::commit();

            // Broadcast proposal status changed event
            $newStatus = $status->value;
            if ($oldStatus !== $newStatus) {
                event(new ProposalStatusChanged($proposal, $oldStatus, $newStatus));
            }

            return ApiResponse::success(
                'Proposal status updated successfully',
                ['proposal' => new ProposalResource($proposal)]
            );
        } catch (\Exception $e) {
            DB::rollBack();

            Log::error('Error updating proposal status', [
                'proposal_id' => $proposal->id,
                'error' => $e->getMessage(),
                'trace' => $e->getTraceAsString(),
            ]);

            return ApiResponse::error('Failed to update proposal status', 500);
        }
    }
}